[2484] Nest access request under providers
